Main file
Note that the comments were a mandatory component of the assignment were meant to clarify what I was doing with each part of the code. The actual file can be found on the one named such; the "notes" file was simply a space for me to try different things before committing to a final product to submit.